requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.1
botocore==1.38.1
certifi==2025.1.31
chardet==5.2.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jmespath==1.0.1
openai==1.14.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
reportlab==4.4.0
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
